silence double -> float truncation warning

+++ UI/SDLGUI.cpp
@@ -346,7 +346,7 @@ void SDLGUI::GLInit()
     // set up perspective with vertical FOV of 50°. 1:1 application
     // window ratio, near plane of 1.0 and far plane of 10.0
     float ratio = Value(m_app_width) * 1.0f / Value(m_app_height);
-    float radians = 50.0f * M_PI / 180.0f;
+    float radians = static_cast<float>(50.0 * M_PI / 180.0);
     float near = 1.0f;
     float far = 10.0f;
     float cotangent = std::cos(radians) / std::sin(radians);
